+++ src/frontend/components/tables/historical.tsx
